+++ example/lib/pages/controls_always_visible_page.dart
@@ -44,7 +44,8 @@ class _ControlsAlwaysVisiblePageState extends State<ControlsAlwaysVisiblePage> {
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 16),
             child: Text(
-              "Controls always visible",
+              "Controls are always visible. Click on button below to"
+              " enable/disable this mode.",
               style: TextStyle(fontSize: 16),
             ),
           ),